+++ src/battle/bullets.h
@@ -36,7 +36,7 @@ extern long lookup(char *name);
 extern char *readFile(char *filename);
 extern float getAngle(int x1, int y1, int x2, int y2);
 extern void addMissileEngineEffect(Bullet *b);
-extern int mod(int n, int x);
+extern float mod(float n, float x);
 extern void addMissileExplosion(Bullet *b);
 extern Entity **getAllEntsWithin(int x, int y, int w, int h, Entity *ignore);
 extern void drawText(int x, int y, int size, int align, SDL_Color c, const char *format, ...);
